Refactored acceptance tests to improve readability of the Given-When-Then (GWT) format
