+++ test/Microsoft.AspNet.WebApi.Acceptance.Tests/Http/Conventions/given a versioned ApiController using conventions/when using a query string and split into two types.cs
@@ -1,0 +1,69 @@
+﻿namespace given_a_versioned_ApiController_using_conventions
+{
+    using FluentAssertions;
+    using Microsoft.Web;
+    using Microsoft.Web.Http.Conventions;
+    using Microsoft.Web.Http.Conventions.Controllers;
+    using System.Collections.Generic;
+    using System.Linq;
+    using System.Net.Http;
+    using System.Threading.Tasks;
+    using Xunit;
+    using static System.Net.HttpStatusCode;
+
+    public class when_using_a_query_string_and_split_into_two_types : ConventionsAcceptanceTest
+    {
+        [Theory]
+        [InlineData( nameof( ValuesController ), "1.0" )]
+        [InlineData( nameof( Values2Controller ), "2.0" )]
+        [InlineData( nameof( Values2Controller ), "3.0" )]
+        public async Task then_get_should_return_200( string controller, string apiVersion )
+        {
+            // arrange
+
+
+            // act
+            var response = await GetAsync( $"api/values?api-version={apiVersion}" ).EnsureSuccessStatusCode();
+            var content = await response.Content.ReadAsAsync<IDictionary<string, string>>();
+
+            // assert
+            response.Headers.GetValues( "api-supported-versions" ).Single().Should().Be( "1.0, 2.0, 3.0" );
+            content.ShouldBeEquivalentTo(
+                new Dictionary<string, string>()
+                {
+                    ["controller"] = controller,
+                    ["version"] = apiVersion
+                } );
+        }
+
+        [Fact]
+        public async Task then_get_should_return_400_for_an_unsupported_version()
+        {
+            // arrange
+
+
+            // act
+            var response = await GetAsync( "api/values?api-version=4.0" );
+            var content = await response.Content.ReadAsAsync<OneApiErrorResponse>();
+
+            // assert
+            response.StatusCode.Should().Be( BadRequest );
+            content.Error.Code.Should().Be( "UnsupportedApiVersion" );
+        }
+
+        [Fact]
+        public async Task then_get_should_return_400_for_an_unspecified_version()
+        {
+            // arrange
+
+
+            // act
+            var response = await GetAsync( "api/values" );
+            var content = await response.Content.ReadAsAsync<OneApiErrorResponse>();
+
+            // assert
+            response.StatusCode.Should().Be( BadRequest );
+            content.Error.Code.Should().Be( "ApiVersionUnspecified" );
+        }
+    }
+}
