--- test/Microsoft.AspNet.WebApi.Acceptance.Tests/OData/Advanced/given/a v3 orders ApiController using attribute routing.cs
@@ -1,40 +1,0 @@
-﻿namespace given
-{
-    using FluentAssertions;
-    using Microsoft.Web;
-    using Microsoft.Web.OData.Advanced;
-    using System;
-    using System.Threading.Tasks;
-    using Xunit;
-
-    public class a_v3_orders_ApiController_using_attribute_routing : AdvancedAcceptanceTest
-    {
-        [Fact]
-        public async Task get_should_return_200()
-        {
-            // arrange
-
-
-            // act
-            var response = await Client.GetAsync( "api/orders?api-version=3.0" );
-            var orders = await response.EnsureSuccessStatusCode().Content.ReadAsExampleAsync( new[] { new { Id = 0, Customer = "" } } );
-
-            // assert
-            orders.ShouldBeEquivalentTo( new[] { new { Id = 1, Customer = "Customer v3.0" } } );
-        }
-
-        [Fact]
-        public async Task get_with_key_should_return_200_without_version()
-        {
-            // arrange
-
-
-            // act
-            var response = await Client.GetAsync( "api/orders/42?api-version=3.0" );
-            var order = await response.EnsureSuccessStatusCode().Content.ReadAsExampleAsync( new { Id = 0, Customer = "" } );
-
-            // assert
-            order.ShouldBeEquivalentTo( new { Id = 42, Customer = "Customer v3.0" } );
-        }
-    }
-}
