--- test/Microsoft.AspNet.WebApi.Acceptance.Tests/OData/Basic/given/a url versioned ODataController split into two types.cs
@@ -1,78 +1,0 @@
-﻿namespace given
-{
-    using FluentAssertions;
-    using Microsoft.Web;
-    using Microsoft.Web.OData.Basic;
-    using System.Linq;
-    using System.Net.Http;
-    using System.Threading.Tasks;
-    using Xunit;
-    using static System.Net.HttpStatusCode;
-
-    public class a_url_versioned_ODataController_split_into_two_types : BasicAcceptanceTest
-    {
-        [Theory]
-        [InlineData( "v1/people" )]
-        [InlineData( "v1/people(42)" )]
-        [InlineData( "v2/people" )]
-        [InlineData( "v2/people(42)" )]
-        [InlineData( "v3/people" )]
-        [InlineData( "v3/people(42)" )]
-        public async Task get_should_return_200( string requestUrl )
-        {
-            // arrange
-
-
-            // act
-            var response = await GetAsync( requestUrl ).EnsureSuccessStatusCode();
-
-            // assert
-            response.Headers.GetValues( "api-supported-versions" ).Single().Should().Be( "1.0, 2.0, 3.0" );
-        }
-
-        [Fact]
-        public async Task patch_should_return_204()
-        {
-            // arrange
-            var person = new { id = 42, firstName = "John", lastName = "Doe", email = "john.doe@somewhere.com" };
-
-            // act
-            var response = await PatchAsync( "v2/people(42)", person );
-
-            // assert
-            response.StatusCode.Should().Be( NoContent );
-        }
-
-        [Theory]
-        [InlineData( "v1/people(42)" )]
-        [InlineData( "v3/people(42)" )]
-        public async Task patch_should_return_405_when_version_could_be_supported( string requestUrl )
-        {
-            // arrange
-            var person = new { id = 42, firstName = "John", lastName = "Doe", email = "john.doe@somewhere.com" };
-
-            // act
-            var response = await PatchAsync( requestUrl, person );
-            var content = await response.Content.ReadAsAsync<OneApiErrorResponse>();
-
-            // assert
-            response.StatusCode.Should().Be( MethodNotAllowed );
-            content.Error.Code.Should().Be( "UnsupportedApiVersion" );
-        }
-
-        [Fact]
-        public async Task patch_should_return_400_when_version_is_unsupported()
-        {
-            // arrange
-            var person = new { id = 42, firstName = "John", lastName = "Doe", email = "john.doe@somewhere.com" };
-
-            // act
-            var response = await PatchAsync( "v4/people(42)", person );
-            var content = await response.Content.ReadAsAsync<OneApiErrorResponse>();
-
-            // assert
-            response.StatusCode.Should().Be( BadRequest );
-            content.Error.Code.Should().Be( "UnsupportedApiVersion" );
-        }
-    }
-}
