--- test/Microsoft.AspNet.WebApi.Acceptance.Tests/OData/Basic/given/a url versioned ODataController.cs
@@ -1,44 +1,0 @@
-﻿namespace given
-{
-    using FluentAssertions;
-    using Microsoft.Web;
-    using Microsoft.Web.OData.Basic;
-    using System.Linq;
-    using System.Net.Http;
-    using System.Threading.Tasks;
-    using Xunit;
-    using static System.Net.HttpStatusCode;
-
-    public class a_url_versioned_ODataController : BasicAcceptanceTest
-    {
-        [Theory]
-        [InlineData( "v1/orders" )]
-        [InlineData( "v1/orders(42)" )]
-        public async Task get_should_return_200( string requestUrl )
-        {
-            // arrange
-
-
-            // act
-            var response = await GetAsync( requestUrl ).EnsureSuccessStatusCode();
-
-            // assert
-            response.Headers.GetValues( "api-supported-versions" ).Single().Should().Be( "1.0" );
-        }
-
-        [Fact]
-        public async Task get_should_return_400_when_version_is_unsupported()
-        {
-            // arrange
-
-
-            // act
-            var response = await GetAsync( "v2/orders" );
-            var content = await response.Content.ReadAsAsync<OneApiErrorResponse>();
-
-            // assert
-            response.StatusCode.Should().Be( BadRequest );
-            content.Error.Code.Should().Be( "UnsupportedApiVersion" );
-        }
-    }
-}
