+++ test/Microsoft.AspNet.WebApi.Acceptance.Tests/OData/Conventions/given a versioned ODataController using conventions/when using a url segment.cs
@@ -1,0 +1,44 @@
+﻿namespace given_a_versioned_ODataController_using_conventions
+{
+    using FluentAssertions;
+    using Microsoft.Web;
+    using Microsoft.Web.OData.Conventions;
+    using System.Linq;
+    using System.Net.Http;
+    using System.Threading.Tasks;
+    using Xunit;
+    using static System.Net.HttpStatusCode;
+
+    public class when_using_a_url_segment : ConventionsAcceptanceTest
+    {
+        [Theory]
+        [InlineData( "v1/orders" )]
+        [InlineData( "v1/orders(42)" )]
+        public async Task then_get_should_return_200( string requestUrl )
+        {
+            // arrange
+
+
+            // act
+            var response = await GetAsync( requestUrl ).EnsureSuccessStatusCode();
+
+            // assert
+            response.Headers.GetValues( "api-supported-versions" ).Single().Should().Be( "1.0" );
+        }
+
+        [Fact]
+        public async Task then_get_should_return_400_for_an_unsupported_version()
+        {
+            // arrange
+
+
+            // act
+            var response = await GetAsync( "v2/orders" );
+            var content = await response.Content.ReadAsAsync<OneApiErrorResponse>();
+
+            // assert
+            response.StatusCode.Should().Be( BadRequest );
+            content.Error.Code.Should().Be( "UnsupportedApiVersion" );
+        }
+    }
+}
