+++ test/Microsoft.AspNetCore.Mvc.Acceptance.Tests/Conventions/given a versioned Controller using conventions/when using a query string and split into two types.cs
@@ -1,0 +1,63 @@
+﻿namespace given_a_versioned_Controller_using_conventions
+{
+    using FluentAssertions;
+    using Microsoft.AspNetCore.Mvc;
+    using Microsoft.AspNetCore.Mvc.Conventions;
+    using Microsoft.AspNetCore.Mvc.Conventions.Controllers;
+    using System.Linq;
+    using System.Net.Http;
+    using System.Threading.Tasks;
+    using Xunit;
+    using static System.Net.HttpStatusCode;
+
+    public class when_using_a_query_string_and_split_into_two_types : ConventionsAcceptanceTest
+    {
+        [Theory]
+        [InlineData( nameof( ValuesController ), "1.0" )]
+        [InlineData( nameof( Values2Controller ), "2.0" )]
+        [InlineData( nameof( Values2Controller ), "3.0" )]
+        public async Task then_get_should_return_200( string controller, string apiVersion )
+        {
+            // arrange
+            var example = new { controller = "", version = "" };
+
+            // act
+            var response = await GetAsync( $"api/values?api-version={apiVersion}" ).EnsureSuccessStatusCode();
+            var content = await response.Content.ReadAsExampleAsync( example );
+
+            // assert
+            response.Headers.GetValues( "api-supported-versions" ).Single().Should().Be( "1.0, 2.0, 3.0" );
+            content.ShouldBeEquivalentTo( new { controller = controller, version = apiVersion } );
+        }
+
+        [Fact]
+        public async Task then_get_should_return_400_for_an_unsupported_version()
+        {
+            // arrange
+
+
+            // act
+            var response = await GetAsync( "api/values?api-version=4.0" );
+            var content = await response.Content.ReadAsAsync<OneApiErrorResponse>();
+
+            // assert
+            response.StatusCode.Should().Be( BadRequest );
+            content.Error.Code.Should().Be( "UnsupportedApiVersion" );
+        }
+
+        [Fact]
+        public async Task then_get_should_return_400_for_an_unspecified_version()
+        {
+            // arrange
+
+
+            // act
+            var response = await GetAsync( "api/values" );
+            var content = await response.Content.ReadAsAsync<OneApiErrorResponse>();
+
+            // assert
+            response.StatusCode.Should().Be( BadRequest );
+            content.Error.Code.Should().Be( "ApiVersionUnspecified" );
+        }
+    }
+}
